refactor(models): move secret field
