+++ web/oauth/models/__init__.py
@@ -1,0 +1,4 @@
+from .config import OAuthClientConfig
+from .secret_name_field import SecretNameField
+
+__all__ = ["OAuthClientConfig", "SecretNameField"]
